--- lib/emis/upload/service.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-module EMIS
-  module Upload
-    class Service < Common::Client::Base
-    end
-  end
-end
