+++ lib/pension_burial/service.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+module PensionBurial
+  class Service < Common::Client::Base
+  end
+end
